Added analysis of first antilac
There are a few things to note about this commit. First, as the
autofluorescencence level was above all three titrations, I wasn't able
to properly compute the fold-change (hence 'pseudo-fold-change') which
is the fold-change in the absence of background subtraction. This
will probably make little difference as there seem to be other issues.

I'll need to do this experimental run again.